Use std::array<int> for _tmpGridSize to read values.  Apply them to _gridSize after validating that they're > 0, and casting the values to size_t

+++ lib/vdc/DCBOV.cpp
@@ -66,7 +66,7 @@ void DCBOV::_InitDimensions()
 {
     _dimsMap.clear();
     std::array<std::string, 3> dimnames = _bovCollection->GetSpatialDimensions();
-    std::array<int, 3>         dimlens = _bovCollection->GetDataSize();
+    std::array<size_t, 3>      dimlens = _bovCollection->GetDataSize();
 
     for (int i = 0; i < dimnames.size(); i++) {
         Dimension dim(dimnames[i], dimlens[i]);
@@ -282,7 +282,7 @@ int DCBOV::_isCoordinateVariable(std::string varname) const
 
 template<class T> void DCBOV::_generateCoordinates(int dim, const vector<size_t> &min, const vector<size_t> &max, T *region) const
 {
-    std::array<int, 3>    dataSize = _bovCollection->GetDataSize();
+    std::array<size_t, 3> dataSize = _bovCollection->GetDataSize();
     std::array<double, 3> origin = _bovCollection->GetBrickOrigin();
     std::array<double, 3> brickSize = _bovCollection->GetBrickSize();
